HW5 in work fix v4
