+++ hw05_parallel_execution/README.md
@@ -1,15 +1,3 @@
-## Описание реализации
-
-В данной задаче я не использовал `done` так как по паттернам он чаще всего должен быть закрыт, как сигнал завершения. В этой задаче `n` воркеров, каждый из которых может послать сигнал остановки работы программы. Т.е. имеем более одного писателя, и чтобы не нарушать принцип закрытия каналов: не закрывать канал, если получатель, и если несколько отправителей, я использую канал `stopCh` с буфером `n`, чтобы не писать лишние `select`.
-
-Логику подсчета кол-ва ошибок вынес в отдельный тип `Stopper`, который также имеет метод, позволяющий писать в канал `stopCh`, сигнализируя генератор об необходимости завершения.
-
-Так как имеем одного отправителя, то я реализовал паттерн через закрытие канала отправителем (только он пишет в этот канал, поэтому принцип закрытия каналов не нарушается).
-
-Добавил тест через `require.Eventually` - оч крутая штука. Для корректной работы запустил Run() в отдельной горутине, чтобы имело смысл проверять число работающих горутин.
-
-Буду рад всем замечаниям. Например, стоит ли доп. вводить `done`-канал? Мне кажется, что нет, так как закрытие `taskStream` - эквивалентно его логике. Правильно ли выносить логику подсчета ошибок в отдельную структуру `Stopper` или это антипаттерн?
-
 ## Домашнее задание №5 «Параллельное исполнение»
 Необходимо написать функцию для параллельного выполнения заданий в n параллельных горутинах:
 * количество создаваемых горутин не должно зависеть от числа заданий, т.е. функция должна запускать n горутин для конкурентной обработки заданий и, возможно, еще несколько вспомогательных горутин;
